+++ bloco7/7.3/ex2-2.js
@@ -1,0 +1,10 @@
+const assert = require('assert');
+// escreva a função wordLengths aqui
+const wordLengths = (words) => words.map((i) => i.length);
+
+const words = ['sun', 'potato', 'roundabout', 'pizza'];
+const expected = [3, 6, 10, 5];
+
+assert.strictEqual(typeof wordLengths, 'function');
+const output = wordLengths(words);
+assert.deepStrictEqual(output, expected);
